requirements.txt
python-dotenv
google-genai
streamlit
fastapi[standard]
uvicorn
pandas
pytesseract
Pillow
opencv-python